refactor: ファイル構造とネーミング規則を統一
PascalCase命名規則の適用とディレクトリ構造の論理的整理により、コードの保守性と可読性を向上。
