--- app/src/main/java/net/t106/sinkerglwallpaper/center_gy.java
@@ -1,78 +1,0 @@
-package net.t106.sinkerglwallpaper;
-
-import android.opengl.GLES32;
-
-/**
- * Center rotating graveyard object for OpenGL ES 3.2
- * Migrated from OpenGL ES 1.0 fixed pipeline
- */
-public class center_gy extends graveyard {
-
-	private float rotation = 0.0f;
-	private static final float ROTATION_SPEED = -0.125f;
-	private static final int MAX_COUNT = 2881;
-
-	public center_gy()
-	{
-		super();
-		// Define quad vertices (same as original)
-		apex = new float[] { -1f, -1f, 1f, -1f, -1f, 1f, 1f, 1f, };
-		coords = new float[] {0.0f, 1.0f, 1.0f, 1.0f, 0.0f, 0.0f, 1.0f, 0.0f, };
-		
-		// Keep legacy buffer creation for compatibility
-		ab = SinkerService.makeFloatBuffer(apex);
-		cb = SinkerService.makeFloatBuffer(coords);
-	}
-	
-	@Override
-	protected void createShaderProgram() {
-		// Use blend shader program for additive blending
-		shaderProgram = ShaderLoader.Programs.createBlendProgram(SinkerService.getContext());
-	}
-	
-	@Override
-	public void Draw(float[] viewMatrix, float[] projectionMatrix) {
-		// Update MVP matrix with current rotation
-		updateMVP(viewMatrix, projectionMatrix);
-		
-		// Bind shader and set uniforms
-		bindShader();
-		
-		// Set texture
-		TextureUtils.bindTexture(0, SinkerService.textures[0]);
-		
-		// Set blend mode to additive (0)
-		ShaderUtils.setUniform1i(blendModeLocation, 0);
-		
-		// Set color (white for no tinting)
-		ShaderUtils.setUniform4f(colorLocation, 1.0f, 1.0f, 1.0f, 1.0f);
-		
-		// Enable blending for additive effect
-		GLES32.glEnable(GLES32.GL_BLEND);
-		GLES32.glBlendFunc(GLES32.GL_ONE, GLES32.GL_ONE);
-		
-		// Bind VAO and draw
-		BufferUtils.bindVAO(vao);
-		BufferUtils.drawQuad();
-		BufferUtils.unbindVAO();
-		
-		// Disable blending
-		GLES32.glDisable(GLES32.GL_BLEND);
-		
-		// Unbind texture
-		TextureUtils.unbindTexture(0);
-	}
-
-	@Override
-	public void Update(float deltaTime) {
-		// Update rotation counter
-		cnt++;
-		if(cnt >= MAX_COUNT) cnt = 0;
-		
-		// Calculate rotation angle
-		rotation = ROTATION_SPEED * cnt;
-		
-		// Update model matrix with rotation
-		modelMatrix = MatrixUtils.rotateZ(rotation);
-	}
-}
